Reordered cell array setup

+++ Model/run.py
@@ -40,7 +40,7 @@ def setup_cells(simulation):
 
     # create the following cell arrays with initial conditions, arrays will default to zero
     simulation.cell_array("locations", override=np.random.rand(simulation.number_cells, 3) * simulation.size)
-    simulation.cell_array("radii", dtype=float, func=lambda: simulation.min_radius)
+    simulation.cell_array("radii", func=lambda: simulation.min_radius)
     simulation.cell_array("motion", dtype=bool, func=lambda: True)
     simulation.cell_array("FGFR", dtype=int, func=lambda: r.randrange(0, simulation.field))
     simulation.cell_array("ERK", dtype=int, func=lambda: r.randrange(0, simulation.field))
@@ -51,8 +51,8 @@ def setup_cells(simulation):
     simulation.cell_array("diff_counters", dtype=int, func=lambda: r.randrange(0, simulation.pluri_to_diff))
     simulation.cell_array("div_counters", dtype=int, func=lambda: r.randrange(0, simulation.pluri_div_thresh))
     simulation.cell_array("fds_counters", dtype=int, func=lambda: r.randrange(0, simulation.fds_thresh))
-    simulation.cell_array("motility_forces", dtype=float, vector=3)
-    simulation.cell_array("jkr_forces", dtype=float, vector=3)
+    simulation.cell_array("motility_forces", vector=3)
+    simulation.cell_array("jkr_forces", vector=3)
     simulation.cell_array("nearest_nanog", dtype=int, func=lambda: -1)
     simulation.cell_array("nearest_gata6", dtype=int, func=lambda: -1)
     simulation.cell_array("nearest_diff", dtype=int, func=lambda: -1)
